Added request to parse_response
This adds the 'request' parameter to the base middleware's
parse_response method. It is intended to support middleware whose
logic is conditional on the request. For example, response header
decorators that respond only to the presence of certain request
headers (If-Not-Match, If-Modified-Since, Origin, etc).

The request parameter has been made optional, in order to be
backwards compatible with older middleware. Additional inspection
has been provided, as well as tests for backwards compatibility.

Change-Id: I5fa94f74765f14bcf79c48ba4c813929da449be3