Deprecated set_latent
With the introduction of set_defaults, as well as new documentation
features for default values in oslo_config, the CORS middleware
has been placed in the odd position where its configuration behaves
inconsistently. A user may use set_defaults to set persisted default
values that may be overridden, or they may use set_latent to add values
that will be applied to every request. As both of these serve essentially
the same function, and the latter is only necessary if it is feasible to
configure multiple allowed_origins with different CORS properties.

With the depecation of multiple configuration blocks, it is no longer
necessary to maintain this feature. Therefore, this patch adds the necessary
deprecation flags, and removes it from the documentation.

Change-Id: Icd44684b3d05ff6a07665348c08adff8245f2523