make sure we handle the forwarded for headers
The X-Forwarded-For header or the "for=" component in the Forwarded
header (per RFC 7239) is essential for applications to keep track of
the originating peer IP. CADF auditing, for example, is relying on
REMOTE_ADDR to accurating keeping track of the real client address.

If either headers are present, the REMOTE_ADDR WSGI request environment
variable will be replaced by the forwarded for address. The Forwarded
header have precedence over the X-Forwarded-For header.

Change-Id: Ib690e3274cd00d8cfa8a15d341ac433555495ce6