Allow compatibility headers for request_id
We discovered that Nova is using a local version of the request_id
header code because it's x-compute-request-id predates
x-openstack-request-id, and things were never migrated after the oslo
code was extracted. In order to let Nova use the base code (which may
become more complicated with inbound request-id setting) we need to
support a compat headers field, which is additional http header names
that should be set with the request-id value.

Change-Id: I18a92f8bf2f85fcf30d32be68bae58f64b135df0