Remove deprecated oslo.subdomain from CORS opts
Defining configuration under cors.* blocks is deprecated. However, we
advertise [cors.subdomain] as a CORS section, and still throw a warning
if that section (that we list) is used in config files.

This was causing warnings in the keystone unit tests output, since the
generated config file does contain both of those listed sections [2].

[1] https://github.com/openstack/oslo.middleware/blob/master/oslo_middleware/cors.py#L200-L205
[2] https://github.com/openstack/keystone/blob/master/etc/keystone.conf.sample#L587-L616

Change-Id: Ice114bf1fdceda7bf122f5fd6b98fc88a85c8ff3