Add new basic auth middleware
This adds a middleware that can be used as an alternative to noauth in
standalone environments.

This middleware uses a password file which supports the Apache htpasswd
syntax. This file is read for every request, so no service restart is
required when changes are made.

The only password digest supported is bcrypt, and the ``bcrypt``
python library is used for password checks since it supports ``$2y$``
prefixed bcrypt passwords as generated by the Apache htpasswd utility.

Adapted from I874783b8ece0eedf27a94dfed5163d0c82f8b9de.

Change-Id: I3d80d86afd063af9fa2f411885dbd0dc65a7bbc7