Run mypy from tox
This avoids the need to duplicate our dependency list in multiple
places and allows us to take advantage of tox's dependency management
infrastructure, to ensure we always get the latest and greatest version
of a package allowed by upper-constraints.

Note that the change brings in stevedore type hints, which necessitates
some small fixes.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Change-Id: I17d8e2d78d95c0d780dbefe084092738962016f0